[wip] refactor file structure to create global styles.
