+++ src/app/(Authentication)/OTPFlow/OTPNewPassword/index.tsx
@@ -9,12 +9,9 @@ import {
 } from '../../../../Components/AuthButton/AuthButton';
 import AuthInput from '../../../../Components/AuthInput/AuthInput';
 import { useSession } from '../../../../context/AuthContext';
-import {
-  ErrorMessageText,
-  TitleText,
-} from '../../../../styles/InputScreenStyles';
 import { SafeArea, ContentContainer } from '../../../../styles/global';
-import { inputScreenStyles } from '../../styles';
+import { inputScreenStyles } from '../../../../styles/inputScreen';
+import { ErrorMessageText, TitleText } from '../../../../styles/textStyles';
 
 export default function OTPNewPassword() {
   const [password, setPassword] = useState<string>('');
