--- src/app/(BottomTabNavigation)/Profile/EditAddress/styles.ts
@@ -1,5 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { colors } from '../../../../styles/colors';
-
-export default StyleSheet.create({});
